+++ R/get_hex.R
@@ -4,10 +4,15 @@ get_hex <- function(branch,
                     add_html=TRUE,
                     verbose=TRUE){
   
+  wrn <- "URL key not found in DESCRIPTION file."
   if("URL" %in% desc::desc_fields()){
     URL <- desc::desc_get_field(key = "URL")
-  } else {
-    wrn <- "URL key not found in DESCRIPTION file."
+    URL <- grep("git",strsplit(URL,",")[[1]],value = TRUE)
+    if(length(URL)==0) {
+      warning(wrn)
+      return(NULL)
+    } 
+  } else { 
     warning(wrn)
     return(NULL)
   }