+++ NEWS.md
@@ -15,6 +15,7 @@
 * Add `no-check-CRAN` arg to `BiocCheck` step to allow using bioc checks for 
   packages already on CRAN.
 * Remove redundant "Install package" step (now handled within Dockerfile).
+* Fix `get_hex` in cases where multiple links in *DESCRIPTION* URL.
 
 # rworkflows 0.99.4